fixed bug #3220: removing glossary from trashbin results in error
git-svn-id: http://svn.ilias.de/svn/ilias/trunk@15623 21b2c9ec-7c21-0410-8b45-9bfb8ed2bfc5

+++ Modules/Glossary/classes/class.ilObjGlossary.php
@@ -952,11 +952,14 @@ class ilObjGlossary extends ilObject
 		}
 
 		// delete terms
-		$terms = $this->getTermList();
-		foreach ($terms as $term)
+		if (!$this->isVirtual())
 		{
-			$term_obj =& new ilGlossaryTerm($term["id"]);
-			$term_obj->delete();
+			$terms = $this->getTermList();
+			foreach ($terms as $term)
+			{
+				$term_obj =& new ilGlossaryTerm($term["id"]);
+				$term_obj->delete();
+			}
 		}
 		
 		// delete glossary data entry
